modify: Detail 페이지 렌더링 최적화, 로그인 세션 유지 코드 디버깅 - Detail 페이지   - 채팅 컴포넌트를 분리   - 상품 수량 설정 컴포넌트 분리
