+++ Todo.md
@@ -1,10 +1,12 @@
+- 써드파티 로그인 기능의 세션 유지 기능 구현(Provider data 로 분기?)
 - Detail 페이지 이미지 클릭 시 상세보기 구현
 - 상품 구매 후 바로 구매이력 들어가면 구매한 상품이 업데이트가 안됨
-- Detail 페이지 렌더링 최적화 필요.
+- Detail 페이지 렌더링 최적화 필요. O
     - 채팅창의 렌더링 별개 적용?
     - 수량 조정 시 전체가 리렌더링됨
 - 장바구니 렌더링 최적화 필요
     - 수량 조정 시 전체가 리렌더링됨
+    - 헤더 전체가 아닌 장바구니 버튼이랑만 연동하기
 - Purchase 페이지 렌더링 최적화 필요
     - 배송 정보 입력값 수정, 장바구니 수정 시 전체가 리렌더링됨
     - 구매버튼 눌러도 전체 리렌더링
